+++ src/aui-layout/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-2073](https://issues.liferay.com/browse/AUI-2073) Set the tab order on Form Builder
 * [AUI-1994](https://issues.liferay.com/browse/AUI-1994) Adds borders as separators on the Form Builder
 * [AUI-2033](https://issues.liferay.com/browse/AUI-2033) Change add/resize column icon